Fixes: fixed XML errors and kotlin errors
@devillierscloete had some missing drawablels, so I've set them to "ic_lock" for now, so that the project can compile and he can find and fix them later.

@ST1026 decided to, for some reason, rewrite and add whole sections of the backend, these committs included errors in the SQL code and in kotlin code. I have fixed them, and the front end team will likely have to adapt to these changes.

+++ app/src/main/java/com/example/republicsavingsapp/SecuritySettings.kt
@@ -57,7 +57,7 @@ class SecuritySettings : Fragment(){
             userName = viewModel.firstName,
             userSurname = viewModel.surname,
             email = viewModel.email,
-            userID = viewModel.username,
+            userID = viewModel.username.toInt(),
             userPassword = viewModel.password,
             currency = viewModel.currency,
             biometricEnabled = viewModel.biometricEnabled
